% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst.R
\name{end}
\alias{end}
\title{End execution time pointer}
\usage{
end(sttime)
}
\arguments{
\item{sttime}{object from start}
}
\value{
system total execution time in segs
}
\description{
Return the total execution time in which the process start running
}
\examples{
init<-start()
(finish<-end(init))
}
