% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst.R
\name{maxScoreIndx}
\alias{maxScoreIndx}
\title{Fetch max affinity value}
\usage{
maxScoreIndx(scorematrix)
}
\arguments{
\item{scorematrix}{object handed}
}
\description{
Return the index of the maximum value in the scorematrix
}
