% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst.R
\name{metricm_size}
\alias{metricm_size}
\title{Calculate Metric Map Size}
\usage{
metricm_size(sm_sparse, sketch_map_size)
}
\arguments{
\item{sm_sparse}{sparse similarity matrix object handed by the user}

\item{sketch_map_size}{parameter handed by the user accordingly to the input graph}
}
\description{
Return the size of output graph for matching the input graph
}
