% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst.R
\name{simat}
\alias{simat}
\title{Select the features from a similarity matrix with the highest confidence level}
\usage{
simat(path, sketch_map_size, scoressm)
}
\arguments{
\item{path}{similarity matrix of class matrix}

\item{sketch_map_size}{size of input graph of class numeric}

\item{scoressm}{calculated affinity scores for the similarity matrix}
}
\value{
list object containing highest affinity candidate pairs for two input graphs
}
\description{
Uses the function sst_pagerank from the package sstr to calculate the highest affinity scores between two graphs described in a input similarity matrix
}
\examples{
library(readr)
dataset <- data.matrix(read.csv(system.file("extdata/foo.csv", package = "sstr")))
scoreSimMa<- data.matrix(read_csv(system.file("extdata/scores2.csv", package = "sstr")))
smap_size=41
(resulted_matrix<-simat(path=dataset,sketch_map_size=smap_size,scoressm=scoreSimMa))
}
